+++ async_rithmic/plants/order.py
@@ -30,7 +30,11 @@ class OrderPlant(BasePlant):
         Fetch extended login details for order management, accounts, trade routes etc
         """
 
-        responses = await self._send_and_recv_immediate(template_id=300)
+        responses = await self._send_and_collect(
+            template_id=300,
+            account_id=None,
+            expected_response=dict(template_id=301),
+        )
         response = self._first(responses)
 
         self.login_info = dict(
@@ -39,7 +43,6 @@ class OrderPlant(BasePlant):
             user_type=response.user_type,
         )
 
-        # Note: when reconnecting, we can't call `_send_and_collect` b/c the background recv task might be blocked
         if self.trade_routes is None:
             self.trade_routes = await self._list_trade_routes()
         if self.accounts is None:
@@ -70,8 +73,9 @@ class OrderPlant(BasePlant):
 
     async def _subscribe_to_updates(self, **kwargs):
         for account in self.accounts:
-            await self._send_and_recv_immediate(
+            await self._send_and_collect(
                 account_id=account.account_id,
+                expected_response=dict(template_id=kwargs["template_id"] + 1),
                 **kwargs
             )